requirements.txt
pandas 
apache-airflow
psycopg2-binary
sqlalchemy
apache-airflow
apache-airflow-providers-postgres
airflow-provider-great-expectations
great-expectations
matplotlib
seaborn
numpy
apache-airflow[google]
pytest
pyyaml